# book-network/src/main/java/com/chis/book/book/BookResponse.java
package com.chis.book.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BookResponse {
    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private String synopsis;
    private String ownerName;
    private byte[] cover;
    private double rate;
    private boolean archived;
    private boolean shareable;
}