# book-network/src/main/java/com/chis/book/feedback/FeedbackRequest.java
package com.chis.book.feedback;

import jakarta.validation.constraints.*;

// I not provide the id bc a Feedback not will be updatable.
public record FeedbackRequest(
        @Positive(message = "200")
        @Min(value = 0, message = "201")
        @Max(value = 5, message = "202")
        Double note,
        @NotNull(message = "203")
        @NotEmpty(message = "203")
        @NotBlank(message = "203")
        String comment,
        @NotNull(message = "204")
        Integer bookId
) {
}