# book-network/src/main/java/com/chis/book/feedback/FeedbackResponse.java
package com.chis.book.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {
    private double note;
    private String comment;
    private boolean ownFeedback;
}